Enable external settings menu
Accessed by the `Ctrl + ,` shortcut. Also removed themes shipped with
the extension in favour of using the CDN.

The original extension settings menu has remain untouched by this
commit and has known compatability issues with the external menu if
used. Otherwise everything should work as before for the casual user.